# MQE: add CSE support for range vector expressions in instant queries