--- a/pkg/streamingpromql/config.go
+++ b/pkg/streamingpromql/config.go
@@ -24,12 +24,14 @@ type EngineOpts struct {
 	// but for now, we use this option to change the behavior of selectors.
 	EagerLoadSelectors bool `yaml:"-"`
 
-	EnableCommonSubexpressionElimination bool `yaml:"enable_common_subexpression_elimination" category:"experimental"`
-	EnableSkippingHistogramDecoding      bool `yaml:"enable_skipping_histogram_decoding" category:"experimental"`
+	EnableCommonSubexpressionElimination                                          bool `yaml:"enable_common_subexpression_elimination" category:"experimental"`
+	EnableCommonSubexpressionEliminationForRangeVectorExpressionsInInstantQueries bool `yaml:"enable_common_subexpression_elimination_for_range_vector_expressions_in_instant_queries" category:"experimental"`
+	EnableSkippingHistogramDecoding                                               bool `yaml:"enable_skipping_histogram_decoding" category:"experimental"`
 }
 
 func (o *EngineOpts) RegisterFlags(f *flag.FlagSet) {
 	f.BoolVar(&o.EnableCommonSubexpressionElimination, "querier.mimir-query-engine.enable-common-subexpression-elimination", true, "Enable common subexpression elimination when evaluating queries.")
+	f.BoolVar(&o.EnableCommonSubexpressionEliminationForRangeVectorExpressionsInInstantQueries, "querier.mimir-query-engine.enable-common-subexpression-elimination-for-range-vector-expressions-in-instant-queries", true, "Enable common subexpression elimination for range vector expressions when evaluating instant queries. Has no effect if common subexpression elimination is disabled.")
 	f.BoolVar(&o.EnableSkippingHistogramDecoding, "querier.mimir-query-engine.enable-skipping-histogram-decoding", true, "Enable skipping decoding native histograms when evaluating queries that do not require full histograms.")
 }
 
@@ -47,7 +49,8 @@ func NewTestEngineOpts() EngineOpts {
 
 		Pedantic: true,
 
-		EnableCommonSubexpressionElimination: true,
-		EnableSkippingHistogramDecoding:      true,
+		EnableCommonSubexpressionElimination:                                          true,
+		EnableCommonSubexpressionEliminationForRangeVectorExpressionsInInstantQueries: true,
+		EnableSkippingHistogramDecoding:                                               true,
 	}
 }