--- a/pkg/streamingpromql/optimize/plan/commonsubexpressionelimination/node.go
+++ b/pkg/streamingpromql/optimize/plan/commonsubexpressionelimination/node.go
@@ -72,20 +72,32 @@ func (d *Duplicate) OperatorFactory(children []types.Operator, _ types.QueryTime
 		return nil, fmt.Errorf("expected exactly 1 child for Duplicate, got %v", len(children))
 	}
 
-	inner, ok := children[0].(types.InstantVectorOperator)
-	if !ok {
+	switch inner := children[0].(type) {
+	case types.InstantVectorOperator:
+		return &InstantVectorDuplicationConsumerOperatorFactory{
+			Buffer: NewInstantVectorDuplicationBuffer(inner, params.MemoryConsumptionTracker),
+		}, nil
+	case types.RangeVectorOperator:
+		return &RangeVectorDuplicationConsumerOperatorFactory{
+			Buffer: NewRangeVectorDuplicationBuffer(inner, params.MemoryConsumptionTracker),
+		}, nil
+	default:
 		return nil, fmt.Errorf("expected InstantVectorOperator as child of Duplicate, got %T", children[0])
 	}
-
-	return &DuplicationConsumerOperatorFactory{
-		Buffer: NewInstantVectorDuplicationBuffer(inner, params.MemoryConsumptionTracker),
-	}, nil
 }
 
-type DuplicationConsumerOperatorFactory struct {
+type InstantVectorDuplicationConsumerOperatorFactory struct {
 	Buffer *InstantVectorDuplicationBuffer
 }
 
-func (d *DuplicationConsumerOperatorFactory) Produce() (types.Operator, error) {
+func (d *InstantVectorDuplicationConsumerOperatorFactory) Produce() (types.Operator, error) {
+	return d.Buffer.AddConsumer(), nil
+}
+
+type RangeVectorDuplicationConsumerOperatorFactory struct {
+	Buffer *RangeVectorDuplicationBuffer
+}
+
+func (d *RangeVectorDuplicationConsumerOperatorFactory) Produce() (types.Operator, error) {
 	return d.Buffer.AddConsumer(), nil
 }