… CHANGELOG.md
… cmd/mimir/config-descriptor.json
… docs/sources/mimir/configure/configuration-parameters/index.md
… pkg/streamingpromql/benchmarks/benchmarks.go
… pkg/streamingpromql/optimize/plan/commonsubexpressionelimination/instant_vector_operator_test.go
… pkg/streamingpromql/optimize/plan/commonsubexpressionelimination/optimization_pass_test.go
… pkg/streamingpromql/optimize/plan/commonsubexpressionelimination/range_vector_operator_test.go
… pkg/streamingpromql/optimize/plan/commonsubexpressionelimination/series_data_ring_buffer.go
… pkg/streamingpromql/optimize/plan/commonsubexpressionelimination/series_data_ring_buffer_test.go
… pkg/streamingpromql/optimize/plan/skip_histogram_decoding_test.go
… pkg/streamingpromql/testdata/ours/common_subexpression_elimination.test
… pkg/streamingpromql/types/hpoint_ring_buffer.go
… pkg/streamingpromql/types/ring_buffer_test.go