#### What this PR does

In https://github.com/grafana/mimir/pull/11189, we added common subexpression elimination (CSE) to MQE's query planner. However, this only supported expressions that produced instant vectors - so this would apply to expressions like `sum(foo)` in `sum(foo) + sum(foo)` or `rate(foo[5m])` in `rate(foo[5m]) + rate(foo[5m])`.

In this PR, I'm adding support for expressions that produce range vectors, provided the parent query is an instant query. For example, CSE will now apply to `foo[5m]` in `sum_over_time(foo[5m]) / count_over_time(foo[5m])`.

I've limited this to instant queries for the time being as supporting range queries adds a lot of complexity. The vast majority of queries we see in production Mimir clusters are instant queries from rules, so we get most of the benefit without the extra complexity.

Benchmarks show a 40-50% latency improvement for affected queries, albeit at the cost of higher peak memory consumption due to the buffering required, similar to what we see for instant vector expressions:

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/streamingpromql/benchmarks
cpu: Apple M1 Pro
                                                                                                        │  before.txt  │             after.txt              │
                                                                                                        │    sec/op    │   sec/op     vs base               │
Query/sum(sum_over_time(a_1[1m]))_/_sum(count_over_time(a_1[1m])),_instant_query/engine=Mimir-10           274.1µ ± 1%   152.4µ ± 1%  -44.41% (p=0.002 n=6)
Query/sum(sum_over_time(a_100[1m]))_/_sum(count_over_time(a_100[1m])),_instant_query/engine=Mimir-10       838.4µ ± 2%   495.7µ ± 0%  -40.87% (p=0.002 n=6)
Query/sum(sum_over_time(a_2000[1m]))_/_sum(count_over_time(a_2000[1m])),_instant_query/engine=Mimir-10     8.555m ± 1%   5.356m ± 0%  -37.40% (p=0.002 n=6)
Query/sum(sum_over_time(a_1[1d]))_/_sum(count_over_time(a_1[1d])),_instant_query/engine=Mimir-10          1136.0µ ± 1%   606.9µ ± 2%  -46.58% (p=0.002 n=6)
Query/sum(sum_over_time(a_100[1d]))_/_sum(count_over_time(a_100[1d])),_instant_query/engine=Mimir-10       68.34m ± 1%   36.16m ± 2%  -47.09% (p=0.002 n=6)
Query/sum(sum_over_time(a_2000[1d]))_/_sum(count_over_time(a_2000[1d])),_instant_query/engine=Mimir-10    1306.9m ± 0%   666.7m ± 1%  -48.98% (p=0.002 n=6)

                                                                                                        │  before.txt  │               after.txt               │
                                                                                                        │      B       │       B        vs base                │
Query/sum(sum_over_time(a_1[1m]))_/_sum(count_over_time(a_1[1m])),_instant_query/engine=Mimir-10          56.76Mi ± 1%    56.53Mi ± 0%         ~ (p=0.223 n=6)
Query/sum(sum_over_time(a_100[1m]))_/_sum(count_over_time(a_100[1m])),_instant_query/engine=Mimir-10      56.73Mi ± 1%    56.52Mi ± 1%         ~ (p=0.509 n=6)
Query/sum(sum_over_time(a_2000[1m]))_/_sum(count_over_time(a_2000[1m])),_instant_query/engine=Mimir-10    58.97Mi ± 1%    59.06Mi ± 1%         ~ (p=0.420 n=6)
Query/sum(sum_over_time(a_1[1d]))_/_sum(count_over_time(a_1[1d])),_instant_query/engine=Mimir-10          62.27Mi ± 1%    62.34Mi ± 1%         ~ (p=0.554 n=6)
Query/sum(sum_over_time(a_100[1d]))_/_sum(count_over_time(a_100[1d])),_instant_query/engine=Mimir-10      69.50Mi ± 3%   116.27Mi ± 3%   +67.30% (p=0.002 n=6)
Query/sum(sum_over_time(a_2000[1d]))_/_sum(count_over_time(a_2000[1d])),_instant_query/engine=Mimir-10    75.62Mi ± 4%   430.17Mi ± 0%  +468.88% (p=0.002 n=6)

```

Benchmarks show no noticable impact for the existing instant vector implementation due to making `SeriesDataRingBuffer` generic.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/11189

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.