@56quarters 2025-07-31
> > The only feedback I have (which might be too vague to be actionable) is that the code for `RangeVectorDuplicationBuffer` involves a lot of operations indexing into slices and checking for specific sentinel values. Maybe it's possible to move `currentSeriesIndex` and `hasReadCurrentSeriesSamples` (and maybe other members) into a separate structure that doesn't require callers to keep track of slice indexes?
> 
> Good point, how's what I've got in [75af8c3](https://github.com/grafana/mimir/commit/75af8c32c58e33172826e0270f0d3aad8220d0ed)?

Looks good. Thanks!